Bring Google Map API call to App.js
As there are multiple features that rely on the google map api, we shall call it in App.js at the top of the component tree, allowing features to rely on the single call made at the beginning.

The isLoaded variable is passed down to the components for error handling should the api call fail.

Map.js was commented out to prevent any bugs